Create Sudoku Game Backtracking Algorithm
This algorithm can be used to built a sudoku game using python pygame.
In this algorithm, a random, solvable board is created and the user can attempt to solve it by clicking on the cells. Entering a number into a cell will be entered as a tentative value. Once the the user is sure that the inputted number is the correct entry, pressing the enter key will attempt to input the number onto the board. Correct answers will be permanently displayed while incorrect answers will be removed.